+++ src/imio/smartweb/policy/upgrades/upgrades.py
@@ -61,3 +61,30 @@ def set_keycloak_login_group(context):
         oidc.allowed_groups = ["iA.Smartweb"]
     else:
         logger.warning("OIDC plugin not found in acl_users; cannot set allowed_groups.")
+
+
+def uninstall_plone_patternslib(context):
+    installer = get_installer(context)
+    installer.uninstall_product("plone.patternslib")
+
+
+def install_plone_patternslib(context):
+    portal_setup = api.portal.get_tool("portal_setup")
+    portal_setup.runAllImportStepsFromProfile("profile-plone.patternslib:default")
+    ps = api.portal.get_tool("portal_setup")
+    PROFILE = "profile-plone.patternslib:default"
+    ps.setLastVersionForProfile(PROFILE, "0")
+    steps = ps.listUpgrades(PROFILE)
+    target_id = ""
+    for step in steps:
+        if (
+            step[0].get("title")
+            == "Update the Patternslib bundle and clean up the old registered"
+        ):
+            target_id = step[0].get("id")
+            break
+    req = context.REQUEST
+    req.form["profile_id"] = PROFILE
+    req.form["upgrades"] = [target_id]
+    ps.manage_doUpgrades(req)
+    ps.setLastVersionForProfile(PROFILE, ps.getLastVersionForProfile(PROFILE))